+++ components/Crumbtrail.vue
@@ -11,7 +11,14 @@
 <script setup>
 
 const props = defineProps({
-    crumbs: Object,
+    crumbs: {
+        type: Array,
+        required: true,
+        validator : (value) => z.object({
+                    label: z.string().min(1),
+                    to: z.string().min(1),
+                }).strict().array().safeParse(value).success,
+    },
 });
 
 </script>